add deserializer for Specializations
