+++ backend/src/main/java/ru/hh/career/solution/parseProfEnvAndSpec/dto/SpecializationsDeserializer.java
@@ -10,23 +10,21 @@ import java.io.IOException;
 import java.util.Arrays;
 
 
-public class SpecializationsDeserializer extends StdDeserializer<ProfessionalEnvironmentDTO> {
+public class SpecializationsDeserializer extends StdDeserializer<SpecializationsDTO> {
     public SpecializationsDeserializer() {
-        super(ProfessionalEnvironmentDTO.class);
+        super(SpecializationsDeserializer.class);
     }
 
     @Override
-    public ProfessionalEnvironmentDTO deserialize(JsonParser jsonParser, DeserializationContext deserializationContext) throws IOException {
-        ProfessionalEnvironmentDTO profEnv = new ProfessionalEnvironmentDTO();
+    public SpecializationsDTO deserialize(JsonParser jsonParser, DeserializationContext deserializationContext) throws IOException {
+        SpecializationsDTO specDTO = new SpecializationsDTO();
         JsonNode treeNode = jsonParser.readValueAsTree();
         if (treeNode == null) {
             return null;
         }
-        profEnv.setId(treeNode.get("id").asInt());
-        profEnv.setName(treeNode.get("name").asText());
-        ObjectMapper objectMapper = new ObjectMapper();
-        SpecializationsDTO[] specDTOS = objectMapper.readValue(treeNode.get("specializations").toString(), SpecializationsDTO[].class);
-        profEnv.setSpecialization(Arrays.asList(specDTOS));
-        return profEnv;
+        specDTO.setId(treeNode.get("id").asInt());
+        specDTO.setName(treeNode.get("name").asText());
+
+        return specDTO;
     }
 }
